requirements.txt
dwave-ocean-sdk>=3.3.0
dwave-neal>=0.4.2
networkx==2.6.2
matplotlib
numpy==1.21.4
opencv-python>=4.5.1.48
PyMaxflow==1.2.13